requirements_javec.txt
flask
flask_marshmallow
pandas
matplotlib
flask_sqlalchemy
python-crontab
flask_restful
click
